edit style of login page #74
